+++ app/models/alchemy/page/page_natures.rb
@@ -2,6 +2,15 @@ module Alchemy
   module Page::PageNatures
     extend ActiveSupport::Concern
 
+    def public?
+      current_time = Time.current
+      already_public_for?(current_time) && still_public_for?(current_time)
+    end
+
+    def expiration_time
+      public_until? ? public_until - Time.current : nil
+    end
+
     def taggable?
       definition['taggable'] == true
     end
@@ -143,6 +152,14 @@ module Alchemy
     def caching_enabled?
       Alchemy::Config.get(:cache_pages) &&
         Rails.application.config.action_controller.perform_caching
+    end
+
+    def already_public_for?(time)
+      !public_on.nil? && public_on <= time
+    end
+
+    def still_public_for?(time)
+      public_until.nil? || public_until >= time
     end
   end
 end